+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha.xx]-2025-02-03
+
+### Added
+
+- Categories API endpoints:
+  - List categories
+  - List nested categories
+  - Get category by id
+  - Get category by slug
+
 ## [1.0.0-alpha.24]-2025-01-31
 
 ### Added